requirements.txt
# TTS相关依赖
soundfile
requests
numpy

# 数据库相关
aiomysql

# Redis相关
redis

# 对象存储相关
tos

# FastAPI相关
fastapi
uvicorn
pydantic

# 其他依赖
python-multipart
python-dotenv

# vllm相关依赖
vllm==0.10.2
descript-audiotools==0.7.2
matplotlib==3.8.2
omegaconf
sentencepiece
librosa
gradio
ninja
modelscope
munch==4.0.0

WeTextProcessing; platform_machine != "Darwin"
wetext; platform_system == "Darwin"